% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{add_clust2enrich}
\alias{add_clust2enrich}
\title{Add clustering assignments to enrichment data frame}
\usage{
add_clust2enrich(enrich, clustdat)
}
\arguments{
\item{enrich}{Preprocessed enrichment data frame (also see function: \link{add_enrich})}

\item{clustdat}{Object of type list storing clustering data (also see function: \link{clust_matrix})}
}
\value{
Processed enrichment data frame with added column indicating cluster assignments
}
\description{
Add clustering assignments to enrichment data frame
}
\examples{
list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
clustering <- clust_matrix(matrix, method = "kmeans", nclust = 3)
enr <- add_enrich(list, '.../enrichment_dat.csv') #file can be pulled from GitHub demo files
enr <- add_clust2enrich(enr, clustering)
}
