% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{add_clust2ts}
\alias{add_clust2ts}
\title{Add clustering assignments to time series data}
\usage{
add_clust2ts(plist, clustdat)
}
\arguments{
\item{plist}{List storing patient time series data (also see function: \link{patient_list})}

\item{clustdat}{Object of type list storing clustering data (also see function: \link{clust_matrix})}
}
\value{
Processed data frame storing time series data with added column indicating cluster assignments
}
\description{
Add clustering assignments to time series data
}
\examples{
list <- patient_list("https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv", GitHub = TRUE)
#Sampling frequency is supposed to be daily
clustering <- clust_matrix(matrix, method = "kmeans", nclust = 3)
ts <- add_clust2ts(list, clustering)

}
