% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust_matrix}
\alias{clust_matrix}
\title{Cluster Earth Mover Distance Square Matrix data and record cluster assignments
for involved Patient_IDs for a specified clustering technique and number of
clusters.}
\usage{
clust_matrix(matrix, method, nclust, plotclust)
}
\arguments{
\item{matrix}{Object of type matrix storing Earth Mover's Distances for patient time series data distribution pairs}

\item{method}{Clustering method (hierarchical, kmeans, diana, fanny, som, modelbased, sota, pam, clara)}

\item{nclust}{Number of clusters (if not specified, user will be asked in the terminal)}

\item{plotclust}{TRUE/FALSE if clustering data should be visualized (TRUE by default)}
}
\value{
Object of type list storing cluster data and clustering assignments for the Patient_IDs from the Earth Mover's Distance matrix
}
\description{
Cluster Earth Mover Distance Square Matrix data and record cluster assignments
for involved Patient_IDs for a specified clustering technique and number of
clusters.
}
\examples{
list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#Sampling frequency is supposed to be daily
matrix <- emd_matrix(list, "FEV1")
clustering <- clust_matrix(matrix, method = "hierarchical", nclust = 2)

}
