% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clValid.R
\name{clValid_flow}
\alias{clValid_flow}
\title{Interactive console workflow to calculate and evaluate cluster validation measures}
\usage{
clValid_flow(matrix, par)
}
\arguments{
\item{matrix}{Earth Mover's Distance Matrix for processed patient time series data (also see functions: \link{emd_matrix}, \link{patient_list})}

\item{par}{Object of type list storing clustering methods and cluster range of interest; initialized via function: \link{init_clValid}}
}
\value{
Object of type list storing chosen clustering method and number of clusters (can be then used for function \link{clust_matrix})
}
\description{
Interactive console workflow to calculate and evaluate cluster validation measures
}
\examples{
list <- patient_list('.../ts_demofiles2') #file can be pulled from GitHub demo files
#(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles2)
#Sampling frequency is twice daily
distmat <- emd_matrix(list, "PEF", maxIter = 5000)
parameters <- init_clValid()
output <- clValid_flow(distmat, parameters)
clustdat <- clust_matrix(distmat, output$method, as.numeric(output$clust_num))

}
