% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{emd_heatmap}
\alias{emd_heatmap}
\title{Visualize an Earth Mover's Distance Square Matrix as a heatmap}
\usage{
emd_heatmap(input, parameter)
}
\arguments{
\item{input}{Earth Mover's Distance Matrix or list storing patient time series data (also see function: \link{patient_list(path)})}

\item{parameter}{In case list is input, the parameter of interest from time series data list}
}
\value{
Visualized Earth Mover's Distance Matrix as a heatmap
}
\description{
Visualize an Earth Mover's Distance Square Matrix as a heatmap
}
\examples{
list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
matrix <- emd_matrix(list, "FEV1")
emd_heatmap(matrix)
}
