% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{emd_matrix}
\alias{emd_matrix}
\title{Generate an Earth Mover's Distance Matrix for time series data distributions
pairs out of a preprocessed time series data list.}
\usage{
emd_matrix(plist, parameter, maxIter)
}
\arguments{
\item{plist}{List storing patient time series data (also see function: \link{patient_list})}

\item{parameter}{Parameter of interest to determine Earth Mover's Distances between distributions}

\item{maxIter}{Maximum of iterations to calculate Earth Mover's Distance (default: 500)}
}
\value{
Earth Mover's Distance Square Matrix of type matrix
}
\description{
Generate an Earth Mover's Distance Matrix for time series data distributions
pairs out of a preprocessed time series data list.
}
\examples{
list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#Sampling frequency is supposed to be daily
matrix <- emd_matrix(list, "FEV1")

}
