% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_and_process.R
\docType{package}
\name{FBCanalysis}
\alias{FBCanalysis}
\title{FBCanalysis: A package for developing and evaluating biomedical time series
data clustering model based on Fluctuation Based Clustering (FBC)}
\description{
This R package aims to offer researchers with fast tools for clustering
patient time series data and confirming the distinction using additional
metrics such as population parameter enrichment analysis, stability after
random data removal, and conventional cluster stability measures.
}
\section{FBCanaylsis functions}{
 \link{add_clust2enrich}

\link{addclust_2ts}

\link{add_enrich}

\link{clust_matrix}

\link{clValid_flow}

\link{emd_heatmap}

\link{emd_matrix}

\link{enr_obs_clust}

\link{addclust_2ts}

\link{init_clValid}

\link{jaccard_run_cognate}

\link{jaccard_run_emd}

\link{max_fluc}

\link{patient_boxplot}

\link{patient_hist}

\link{patient_list}

\link{patient_ts_plot}

\link{rnd_dat_rm}

\link{addclust_2ts}

\link{sim_jaccard_cognate}

\link{sim_jaccard_emd}

\link{sim_sample_enr}

\link{znorm}
}

\seealso{
\href{https://github.com/MrMaximumMax/FBCanalysis}{GitHub Repository}
}
