% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clValid.R
\name{init_clValid}
\alias{init_clValid}
\title{Initialize Cluster Validation Measure Analysis in the context of Fluctuation
Based Clustering (FBC) analysis. The call initialized a interactive console
workflow where the user may indicate the clustering techniques of interest
as well as the cluster numbers of interest.}
\usage{
init_clValid()
}
\value{
Object of type list storing cluster method(s) and number of cluster range of interest (to be used for function: \link{clValid_flow})
}
\description{
Initialize Cluster Validation Measure Analysis in the context of Fluctuation
Based Clustering (FBC) analysis. The call initialized a interactive console
workflow where the user may indicate the clustering techniques of interest
as well as the cluster numbers of interest.
}
\examples{
init_clValid()

}
