% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard_run_cognate}
\alias{jaccard_run_cognate}
\title{Simulate random data removal from time series data list and determine Jaccard index via Cognate Cluster approach for multiple random data removal steps}
\usage{
jaccard_run_cognate(
  plist,
  parameter,
  n_simu,
  method,
  clust_num,
  n_clust,
  range
)
}
\arguments{
\item{plist}{Object of type list storing patient time series data (also see function: \link{patient_list})}

\item{parameter}{Parameter of interest in time series data list}

\item{n_simu}{Number of simulations}

\item{method}{Clustering method (also see function: \link{clust_matrix})}

\item{clust_num}{Cluster of interest}

\item{n_clust}{Number of clusters}

\item{range}{Range to simulate random data removal (e.g. c(0.1,0.2,0.5,0.7,0.8))}
}
\value{
Object of type list storing Jaccard indices for each indicated random data removal step and visualized results in a boxplot
}
\description{
Simulate random data removal from time series data list and determine Jaccard index via Cognate Cluster approach for multiple random data removal steps
}
\examples{
list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
#(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
output <- jaccard_run_cognate(list,"PEF",10,"hierarchical",1,3,c(0.005,0.01,0.05,0.1,0.2))
}
