% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard_run_emd_2}
\alias{jaccard_run_emd_2}
\title{Simulate random data removal range via alternative Earth Mover's Distance
Cognate Cluster cluster assignment approach}
\usage{
jaccard_run_emd_2(
  plist,
  parameter,
  n_simu,
  method,
  clust_num,
  n_clust,
  range,
  maxIter,
  normalize
)
}
\arguments{
\item{plist}{Object of type list storing patient time series data (also see function: \link{patient_list})}

\item{parameter}{Parameter of interest in time series data list}

\item{n_simu}{Number of simulations}

\item{method}{Clustering method (also see function: \link{clust_matrix})}

\item{clust_num}{Cluster of interest}

\item{n_clust}{Number of clusters}

\item{range}{Range to simulate random data removal (e.g. c(0.1,0.2,0.5,0.7,0.8))}

\item{maxIter}{Maximum iterations to determine Earth Mover's Distances (also see function: \link{emd_matrix});
default is 5,000 for this function}

\item{normalize}{Indicates if parameter indicated needs to be normalized or not (TRUE by default)}
}
\value{
Object of type list storing Jaccard indices for each indicated random data
removal step and visualized results in a boxplot
}
\description{
Alternative: Simulate amount of random data removal from time series data list
and determine Jaccard index via Earth Mover's Distance approach for multiple
random data removal steps for a specific cluster of interest.
}
\details{
See \link{sim_jaccard_emd_2} for more detailed approach on Jaccard
index determination. The difference in this function is that now only one cluster
is observed für multiple amounts of random data removal where for each data
removal step defined the resulting Jaccard indices are stored in a list object.
Furthermore, a boxplot visualization is generated, in the style of recent
publications.
}
\examples{
list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#Sampling frequency is supposed to be daily
output <- jaccard_run_emd_2(list,"PEF",10,"hierarchical",1,3,c(0.005,0.01,0.05,0.1,0.2))


}
\references{
Yossi Rubner, Carlo Tomasi, and Leonidas J Guibas. A metric for
distributions with applications to image databases. In Sixth International
Conference on Computer Vision (IEEE Cat. No. 98CH36271), pages 59–66. IEEE, 1998.
}
