% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_and_process.R
\name{patient_boxplot}
\alias{patient_boxplot}
\title{Visualize patient(s) time series data in a boxplot for indicated parameter}
\usage{
patient_boxplot(plist, patients, parameter, normalized)
}
\arguments{
\item{plist}{List storing patient time series data (also see function: \link{patient_list(path)})}

\item{patients}{Patient_ID(s) referring to (a) list element; can be single ID or multiple IDs (also see function: \link{patient_list(path)})}

\item{parameter}{Parameter of interest in list element(s)}

\item{normalized}{TRUE/FALSE if z-normalized (TRUE by default)}
}
\value{
Visualized patient(s) time series data in a boxplot for indicated parameter
}
\description{
Visualize patient(s) time series data in a boxplot for indicated parameter
}
\examples{
list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
patient_boxplot(list,c("ID_2","testpat_1","testpat_2","a301"), "FEV1")
}
