% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_and_process.R
\name{patient_ts_plot}
\alias{patient_ts_plot}
\title{Visualize patient data in time series plot}
\usage{
patient_ts_plot(plist, Patient_ID, parameter, normalize)
}
\arguments{
\item{plist}{List storing patient time series data (also see function: \link{patient_list})}

\item{Patient_ID}{Patient_ID referring to a list element (also see function: \link{patient_list})}

\item{parameter}{Parameter of interest in list element}

\item{normalize}{TRUE/FALSE if z-normalization is required (TRUE by default)}
}
\value{
Visualized patient time series data in a time series plot for indicated parameter
}
\description{
Visualize patient time series data from a preprocessed in a time series plot
for an indicated parameter, either as normalized or non-normalized.
}
\examples{
list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#Sampling frequency is supposed to be daily
patient_ts_plot(list,"testpat_1","PEF")

}
