% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{reap_freq_run}
\alias{reap_freq_run}
\title{Determine reappearance frequency for range}
\usage{
reap_freq_run(
  plist,
  parameter,
  n_simu,
  method,
  n_clust,
  range,
  Iter,
  normalize
)
}
\arguments{
\item{plist}{Object of type list storing patient time series data (also see function: \link{patient_list})}

\item{parameter}{Parameter of interest in time series data list}

\item{n_simu}{Number of simulations}

\item{method}{Clustering method (also see function: \link{clust_matrix})}

\item{n_clust}{Number of clusters}

\item{range}{Range to simulate random data removal (e.g. c(0.1,0.2,0.5,0.7,0.8))}

\item{normalize}{Indicates if parameter indicated needs to be normalized or not (TRUE by default)}
}
\value{
List storing vectors with reappearance frequencies for each simulation run
for each removal amount indicated as well as illustrated results in a boxplot
}
\description{
Determine the reappearance frequency for clustered elements in perturbed data
for specified amounts of random data removal
}
\details{
The procedure on determining the reappearance frequenncy is described
in \link{reap_freq}.
}
\references{
Edgar Delgado-Eckert, Oliver Fuchs, Nitin Kumar, Juha Pekkanen,
Jean-Charles Dalphin, Josef Riedler, Roger Lauener, Michael Kabesch, Maciej Kupczyk,
Sven-Erik Dahlen, et al. Functional phenotypes determined by fluctuation-based
clustering of lung function measurements in healthy and asthmatic cohort participants.
Thorax, 73(2):107–115, 2018.
}
