% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{sim_jaccard_emd}
\alias{sim_jaccard_emd}
\title{Simulate random data removal from time series data list and determine Jaccard index via Earth Mover's Distance approach}
\usage{
sim_jaccard_emd(plist, parameter, removal, n_simu, method, n_clust, Iter)
}
\arguments{
\item{plist}{Object of type list storing patient time series data (also see function: \link{patient_list})}

\item{parameter}{Parameter of interest in time series data list}

\item{removal}{Amount of random data removal to determine Jaccard index}

\item{n_simu}{Number of simulations}

\item{method}{Clustering method (also see function: \link{clust_matrix})}

\item{n_clust}{Number of clusters (also see function: \link{clust_matrix})}

\item{Iter}{Maximum iterations to determine Earth Mover's Distances (also see function: \link{emd_matrix}); default is 5,000 for this function}
}
\value{
Object of type matrix storing received Jaccard indices for indicated amount of random data removal for all clusters
}
\description{
Simulate random data removal from time series data list and determine Jaccard index via Earth Mover's Distance approach
}
\examples{
list <- patient_list("https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv", GitHub = TRUE)
output <- sim_jaccard_emd(list, "PEF", 0.05, 10, "hierarchical", 2, 100)

}
