% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{sim_sample_enr}
\alias{sim_sample_enr}
\title{Simulate random sampling for NA entries in enrichment data and check stability of resulting p-values for the enrichment parameters}
\usage{
sim_sample_enr(plist, path, clustdat, clustno, n_sim)
}
\arguments{
\item{plist}{List storing patient time series data (also see function: \link{patient_list})}

\item{path}{Path where enrichment csv file is stored}

\item{clustdat}{Object of type list storing clustering data (also see function: \link{clust_matrix})}

\item{clustno}{Cluster number of interest}

\item{n_sim}{Number of simulations}
}
\value{
Object of type list storing the received p-values for each parameter in a vector and boxplot visualizing the received p-values
}
\description{
Simulate random sampling for NA entries in enrichment data and check stability of resulting p-values for the enrichment parameters
}
\examples{
list <- patient_list("https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv", GitHub = TRUE)
#Sampling frequency is supposed to be daily
path <- 'https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/enrich/enrichment.csv'
test <- sim_sample_enr(list,path,clustering,1,100)
sim_sample_enr <- function(plist, path, clustdat, clustno, n_sim)

}
