% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/znorm.R
\name{znorm}
\alias{znorm}
\title{z-normalise data}
\usage{
znorm(data)
}
\arguments{
\item{data}{Numeric distribution; may be stored in an object of type vector, matrix or data frame}
}
\value{
z-normalized distribution of input
}
\description{
Function applicable on any numeric distribution of data and z-normalize them.
}
\examples{
random_distribution <- runif(n = 50, min = 1, max = 10)
znorm_distribution <- znorm(random_distribution)

}
