% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{add_clust2enrich}
\alias{add_clust2enrich}
\title{Add clustering assignments from clustering output to preprocessed enrichment
data frame.}
\usage{
add_clust2enrich(enrich, clustdat)
}
\arguments{
\item{enrich}{Preprocessed enrichment data frame (also see function: \link{add_enrich})}

\item{clustdat}{Object of type list storing clustering data (also see function: \link{clust_matrix})}
}
\value{
Processed enrichment data frame with added column indicating cluster assignments
}
\description{
Add clustering assignments from clustering output to preprocessed enrichment
data frame.
}
\examples{
list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#Sampling frequency is supposed to be daily
clustering <- clust_matrix(matrix, method = "kmeans", nclust = 3)
enr <- add_enrich(list,
'https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/enrich/enrichment.csv')
enr <- add_clust2enrich(enr, clustering)

}
