% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{add_enrich}
\alias{add_enrich}
\title{Add enrichment data and preprocess for analysis}
\usage{
add_enrich(plist, path)
}
\arguments{
\item{plist}{List storing patient time series data (also see function: \link{patient_list})}

\item{path}{Path where enrichment csv file is stored}
}
\value{
Processed data as object of type data frame; Enrichment data Patient_IDs are matched with Time Series Data List Patient IDs; In case it was indicated, NA values in the enrichment data are filled up by random sampling
}
\description{
Add enrichment data and preprocess for analysis
}
\examples{
list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
#(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
enr <- add_enrich(list, '.../enrichment_dat.csv') #file can be pulled from GitHub demo files
}
