% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust_matrix}
\alias{clust_matrix}
\title{Cluster Earth Mover Distance Square Matrix data}
\usage{
clust_matrix(matrix, method, nclust, plotclust)
}
\arguments{
\item{matrix}{Object of type matrix storing Earth Mover's Distances for patient time series data distribution pairs}

\item{method}{Clustering method (hierarchical, kmeans, diana, fanny, som, modelbased, sota, pam, clara)}

\item{nclust}{Number of clusters (if not specified, user will be asked in the terminal)}

\item{plotclust}{TRUE/FALSE if clustering data should be visualized (TRUE by default)}
}
\value{
Object of type list storing cluster data and clustering assignments for the Patient_IDs from the Earth Mover's Distance matrix
}
\description{
Cluster Earth Mover Distance Square Matrix data
}
\examples{
list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
matrix <- emd_matrix(list, "FEV1")
clustering <- clust_matrix(matrix, method = "hierarchical", nclust = 2)
}
