% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{emd_heatmap}
\alias{emd_heatmap}
\title{Visualize an Earth Mover's Distance Square Matrix as a heatmap}
\usage{
emd_heatmap(input, parameter, maxIter, normalize)
}
\arguments{
\item{input}{Earth Mover's Distance Matrix or list storing patient time series data (also see function: \link{patient_list})}

\item{parameter}{In case list is input, the parameter of interest from time series data list}

\item{normalize}{Indicates if parameter indicated needs to be normalized or not (TRUE by default)}

\item{Iter}{In case input is time series data list, incate maxIter to calculate EMD matrix (also see function: \link{emd_matrix})}
}
\value{
Visualized Earth Mover's Distance Matrix as a heatmap
}
\description{
Visualize an Earth Mover's Distance Square Matrix as a heatmap
}
\examples{
list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#Sampling frequency is supposed to be daily
matrix <- emd_matrix(list, "FEV1")
emd_heatmap(matrix)

}
