% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{emd_matrix}
\alias{emd_matrix}
\title{Generate a Earth Mover's Distance Matrix out of time series data list entries}
\usage{
emd_matrix(plist, parameter, maxIter)
}
\arguments{
\item{plist}{List storing patient time series data (also see function: \link{patient_list(path)})}

\item{parameter}{Parameter of interest to determine Earth Mover's Distances between distributions}

\item{maxIter}{Maximum of iterations to calculate Earth Mover's Distance (default: 500)}
}
\value{
Earth Mover's Distance Square Matrix of type matrix
}
\description{
Generate a Earth Mover's Distance Matrix out of time series data list entries
}
\examples{
list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
matrix <- emd_matrix(list, "FEV1")
}
