% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enr_obs_clust}
\alias{enr_obs_clust}
\title{Observe specific cluster for overview and p-values}
\usage{
enr_obs_clust(ts.dat, enrich, clustno)
}
\arguments{
\item{ts.dat}{Processed data frame storing time series data and cluster assignments (also see function: \link{add_clust2ts})}

\item{enrich}{Processed data frame storing enrichment data and cluster assignments (also see function: \link{add_clust2enrich})}

\item{clustno}{Cluster number of interest}
}
\value{
Terminal output presenting summary of time series and enrichment data with corresponding p-values
}
\description{
Observe specific cluster for overview and p-values
}
\examples{
list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
#(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
clustering <- clust_matrix(matrix, method = "kmeans", nclust = 3)
enr <- add_enrich(list, '.../enrichment_dat.csv') #file can be drawn from GitHub demo files
#(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/enrichment)
enr <- add_clust2enrich(enr, clustering)
ts <- add_clust2ts(list, clustering)
enr_obs_clust(ts, enr, 1)

}
