% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{max_fluc}
\alias{max_fluc}
\title{Determine pair of maximum fluctuation difference in a list storing time series data}
\usage{
max_fluc(plist, parameter)
}
\arguments{
\item{plist}{List storing patient time series data (also see function: \link{patient_list(path)})}

\item{parameter}{Parameter of interest from time series data list}
}
\value{
Console output with Patient_ID pair, corresponding Earth Mover's Distance and visualized boxplot of both time series data distributions
}
\description{
Determine pair of maximum fluctuation difference in a list storing time series data
}
\examples{
list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
#(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
matrix <- emd_matrix(list, "FEV1")
max_fluc(list, "PEF")
}
