% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_and_process.R
\name{patient_list}
\alias{patient_list}
\title{Process patient time series data by interpolation options and store data in an object of type list}
\usage{
patient_list(path, GitHub)
}
\arguments{
\item{path}{Path where csv file(s) are stored (only folder, not specific file(s))}

\item{GitHub}{Set TRUE when csv file comes form GitHub (FALSE by default)}
}
\value{
Object of type list storing patient time series data
}
\description{
Process patient time series data by interpolation options and store data in an object of type list
}
\examples{
list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#Sampling frequency is supposed to be daily

}
