% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_and_process.R
\name{patient_ts_plot}
\alias{patient_ts_plot}
\title{Visualize patient time series data in a time series plot for indicated parameter}
\usage{
patient_ts_plot(plist, Patient_ID, parameter, normalized)
}
\arguments{
\item{plist}{List storing patient time series data (also see function: \link{patient_list})}

\item{Patient_ID}{Patient_ID referring to a list element (also see function: \link{patient_list})}

\item{parameter}{Parameter of interest in list element}

\item{normalized}{TRUE/FALSE if z-normalized (TRUE by default)}
}
\value{
Visualized patient time series data in a time series plot for indicated parameter
}
\description{
Visualize patient time series data in a time series plot for indicated parameter
}
\examples{
list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
#(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
patient_ts_plot(list,"testpat_1","PEF")

}
