% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{rnd_dat_rm}
\alias{rnd_dat_rm}
\title{Remove random data from time series data list}
\usage{
rnd_dat_rm(plist, removal)
}
\arguments{
\item{plist}{Object of type list storing patient time series data (also see function: \link{patient_list})}

\item{removal}{Amount of data removal (0 = 0\%, 1 = 100\%)}
}
\value{
Object of type list storing patient time series data with indicated amount of data removed randomly
}
\description{
Remove random data from time series data list
}
\examples{
list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#Sampling frequency is supposed to be daily
list_rm <- rnd_dat_rm(testlist, 0.95)

}
